refactor: move schedule management to useCreateSchedule
Moves schedule management logic from its own hook to useCreateSchedule.
