+++ lib/hooks/use-create-schedule.ts
@@ -1,11 +1,14 @@
 'use client';
 
-import { useState } from 'react';
+import { useState, useMemo, useCallback } from 'react';
 import { useCourses } from '@/lib/hooks/use-courses';
-import { useScheduleManagement } from '@/lib/hooks/use-schedule-management';
 import { Course } from '@/lib/types/course';
 import { saveSchedule } from '@/lib/services/schedules';
 import { useMessage } from '@/lib/hooks/use-message';
+import {
+  detectTimeConflicts,
+  calculateScheduleStats,
+} from '@/lib/schedule-utils';
 
 export const useCreateSchedule = () => {
   const [activeTab, setActiveTab] = useState('manual');
@@ -15,23 +18,62 @@ export const useCreateSchedule = () => {
 
   const { courses, isLoading } = useCourses();
 
-  const {
-    selectedCourses,
-    searchQuery,
-    filterSemester,
-    filterClass,
-    groupByCode,
-    conflicts,
-    stats,
-    filterCourses,
-    toggleCourse,
-    clearAllSelections,
-    setSelectedCoursesDirectly,
-    setSearchQuery,
-    setFilterSemester,
-    setFilterClass,
-    setGroupByCode,
-  } = useScheduleManagement();
+  // --- Merged from useScheduleManagement ---
+  const [selectedCourses, setSelectedCourses] = useState<Course[]>([]);
+  const [searchQuery, setSearchQuery] = useState('');
+  const [filterSemester, setFilterSemester] = useState('all');
+  const [filterClass, setFilterClass] = useState('all');
+  const [groupByCode, setGroupByCode] = useState(false);
+
+  const conflicts = useMemo(
+    () => detectTimeConflicts(selectedCourses),
+    [selectedCourses]
+  );
+  const stats = useMemo(
+    () => calculateScheduleStats(selectedCourses),
+    [selectedCourses]
+  );
+
+  const filterCourses = useCallback(
+    (coursesToFilter: Course[]) => {
+      return coursesToFilter.filter((course) => {
+        const matchesSearch =
+          searchQuery === '' ||
+          course.name.toLowerCase().includes(searchQuery.toLowerCase()) ||
+          course.code.toLowerCase().includes(searchQuery.toLowerCase()) ||
+          course.lecturer.toLowerCase().includes(searchQuery.toLowerCase());
+
+        const matchesSemester =
+          filterSemester === 'all' || course.semester === filterSemester;
+
+        const matchesClass =
+          filterClass === 'all' || course.class === filterClass;
+
+        return matchesSearch && matchesSemester && matchesClass;
+      });
+    },
+    [searchQuery, filterSemester, filterClass]
+  );
+
+  const toggleCourse = useCallback((course: Course) => {
+    setSelectedCourses((prev) => {
+      const isSelected = prev.find((c) => c.id === course.id);
+      if (isSelected) {
+        return prev.filter((c) => c.id !== course.id);
+      } else {
+        return [...prev, course];
+      }
+    });
+  }, []);
+
+  const clearAllSelections = useCallback(() => {
+    setSelectedCourses([]);
+  }, []);
+
+  const setSelectedCoursesDirectly = useCallback((courses: Course[]) => {
+    setSelectedCourses(courses);
+  }, []);
+  // --- End of merge ---
 
   const filteredCourses = filterCourses(courses);
